+++ packedForest/test/test.cpp
@@ -16,9 +16,9 @@
 #include "fpTests/newForest/processingNodeTest.h"
 #include "fpTests/baseUnprocessedNodeTest.h"
 #include "fpTests/RerF/unprocessedRerFNodeTest.h"
-#include "fpTests/MWC.h"
-#include "fpTests/fpInfo.h"
+#include "fpTests/MWCTest.h"
+#include "fpTests/fpInfoTest.h"
+#include "fpTests/fpForestClassificationBaseTest.h"
+#include "fpTests/fpRerFBaseTest.h"
+#include "fpTests/binnedBaseTest.h"
 #include "fpTests/fpForestTest.h"
-#include "fpTests/fpForestClassificationBase.h"
-#include "fpTests/fpRerFBase.h"
-#include "fpTests/binnedBase.h"
